+++ addons/blender/2.80/io_scene_wowobj/__init__.py
@@ -21,7 +21,7 @@
 bl_info = {
     'name': 'Import WoW OBJ files with doodads',
     'author': 'Marlamin, Kruithne',
-    'version': (0, 3, 10),
+    'version': (0, 3, 11),
     'blender': (2, 92, 0),
     'location': 'File > Import-Export > WoW M2/WMO/ADT (.obj)',
     'description': 'Import OBJ files exported by wow.export with WMOs and doodads',