+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.1.30 (In Development)
 - Added 'Export Textures' option to sidebar to allow models to be exported without textures.
+- Fixed issue that caused incorrect nesting when imported with the Blender add-on.
 
 0.1.29 (29-03-2021)
 - Added new 'Items' tab, allowing model/texture look-up of specific game items.